style: 페이지 공통 layout 적용
